Move dimension/output mapping to dedicated mapper mod.
